Moved placeholders to media folder. Added placeholder images to News Artticales
