+++ src/components/FeatureList/scrollMagicEnhanced.js
@@ -122,7 +122,7 @@ function withSubscription(WrappedComponent, selectData) {
            timeline.play();
           }.bind(this))
           .on("leave", function (event) {
-            timeline.reverse()
+          //  timeline.reverse()
           }.bind(this))
           .addTo(this.controller));
