+++ src/components/FeatureList/styles.scss
@@ -28,21 +28,22 @@
         transform: translateX(150px);
     }
 
-    &:nth-child(odd){
-        @include media-breakpoint-up(md) {
+    @include media-breakpoint-up(md) {
+        &:nth-child(odd){
             text-align: right;
             img {
                 float: right;
-                width: 50px;
-                height: auto;
                 transform: translateX(-50px);
             }
             .text {
-                display: block;
                 padding-left: 0;
                 padding-right: 80px;
                 transform: translateX(-150px);
             }
         }
     }
+
+
+
+
 }
